// shared/api/fine/model.ts
export interface GetFineByNumberApiResponse {
  number: string;
  name: string;
  payee_inn: string;
  payee_kpp: string;
  doc_number: string;
  violation_at: string;
  koap_code: string;
  payee_username: string;
  payee_bank_account: string;
  payee_bank_name: string;
  payee_bank_bik: string;
  payee_oktmo: string;
  kbk: string;
  discount_size: string;
  discount_at: string;
  amount: string;
  bill_at: string;
  amount_to_pay: string;
}